path routes merged to controller
