+++ controllers/paths.js
@@ -5,47 +5,50 @@ module.exports = {
   create,
   show,
   destroy,
+  search,
   indexBlip,
   createBlip,
   showBlip,
   destroyBlip,
-  updateBlip
+  updateBlip,
+  searchPathByBlips
 }
 
+// show all paths
 function index(req, res) {
-  Path.find({}, function(err, paths) {
-    if(err) return console.log(err)
-    res.json(paths)
-  })
-}
+    Path.find({}).populate('_by').exec(function(err, paths) {
+      if(err) return console.log(err)
+      res.render('paths', {paths})
+    })
+  }
 
+// make a new path
 function create(req, res){
-  // This finds the current user by their id
-  User.find(currentUser.id , function(err, user){
-    // creates a new unsaved path with the body in the request
-    var newPath = new Path(req.body)
-    // gives that path a _by property of the user's id
-    newPath._by = user
-    // saves that path
-    newPath.save(function(err, path){
-      // pushes this path id in the user object's paths property
-      user.paths.push(path)
-      // then saves the current user object
-      user.save(function(err, user){
-        // and responds to the client with the user object
-        res.json(user)
+    User.find(currentUser.id , function(err, user){
+      var newPath = new Path(req.body)
+      newPath._by = user
+      newPath.save(function(err, path){
+        user.paths.push(path)
+        user.save(function(err, user){
+          res.json(user)
+        })
       })
     })
-  })
-}
+  }
 
-function show(req, res) {
-  Path.findById(req.params.id).populate('blips').exec(function(err, path) {
-    if(err) return console.log(err)
-    res.json(path)
-  })
-}
+// show a specific path
+function show(req, res){
+    Path.findById(req.params.id).populate('_by').exec(function(err, path) {
+      if(err) return console.log(err)
+      var blips = path.blips.concat()
+      blips = blips.sort(function(obj1, obj2) {
+        return obj2.year - obj1.year
+      })
+      res.render('path', {path, blips})
+    })
+  }
 
+// delete a path
 function destroy(req, res){
     Path.findByIdAndRemove(req.params.id, function(err){
       if(err) return console.log(err)
@@ -59,6 +62,19 @@ function destroy(req, res){
     })
 }
 
+// Search a path
+function search(req, res){
+  // converts search term to a regex object / "i" allows case insensitivity
+  var rx = new RegExp(req.body.name, "i")
+    // finds a path with a name that matches the search term
+    Path.find({"name": rx}, function(err,data){
+      if(err) return res.json(err)
+      // returns that data to the client as a json object
+      res.json(data)
+    })
+  }
+
+// show all blips
 function indexBlip(req, res) {
   Path.findById(req.params.id, function(err, path) {
     if(err) return console.log(err)
@@ -66,17 +82,19 @@ function indexBlip(req, res) {
   })
 }
 
+// make a new blip
 function createBlip(req, res) {
-  Path.findById(req.params.id, function(err, path) {
-    if(err) return console.log(err)
-    path.blips.push(req.body)
-    path.save(function(err) {
+    Path.findById(req.params.id, function(err, path){
       if(err) return console.log(err)
-      res.json(path)
+      path.blips.push(req.body)
+      path.save(function(err) {
+        if(err) return console.log(err)
+        res.json(path)
+      })
     })
-  })
-}
+  }
 
+// show a specific blip
 function showBlip(req, res) {
   Path.findById(req.params.pathId, function(err, path) {
     if(err) return console.log(err)
@@ -84,30 +102,43 @@ function showBlip(req, res) {
   })
 }
 
+// delete a specific blip
 function destroyBlip(req, res) {
   Path.findById(req.params.pathId, function(err, path) {
-    if(err) return console.log(err)
-    path.blips.id(req.params.blipId).remove()
-    path.save(function(err) {
-      if(err) return console.log(err)
-      res.json(path)
-    })
-  })
-}
-
-function updateBlip(req, res){
-  Path.findById(req.params.pathId, function(err, path) {
-    // res.json(path.blips)
-    if(err) return res.json(err)
-    // // path.blips.id(req.params.blipId).update()
-    path.blips.forEach(function(blip){
-      if (blip._id == req.params.blipId){
-        blip.title = req.body.title
-        path.save(function(err){
-          if(err) return res.json(err)
+      Path.findById(req.params.pathId, function(err, path) {
+        if(err) return console.log(err)
+        path.blips.id(req.params.blipId).remove()
+        path.save(function(err) {
+          if(err) return console.log(err)
           res.json(path)
         })
-      }
+      })
     })
-  })
-}
+  }
+
+// update a specific blip
+function updateBlip(req, res){
+    Path.findById(req.params.pathId, function(err, path) {
+      // res.json(path.blips)
+      if(err) return res.json(err)
+      // // path.blips.id(req.params.blipId).update()
+      path.blips.forEach(function(blip){
+        if (blip._id == req.params.blipId){
+          blip.title = req.body.title
+          path.save(function(err){
+            if(err) return res.json(err)
+            res.json(path)
+          })
+        }
+      })
+    })
+  }
+
+// search for a path by its blips
+function searchPathByBlips(req, res) {
+  var rx = new RegExp(req.body.title, "i")
+    // looks within all paths for blips whose name matches the search term
+    Path.find({"blips.title": rx}, function(err,paths){
+      res.json(paths)
+    })
+  }
